Split into two components (#333)
