+++ apcd-cms/src/client/src/components/Submitter/Extensions/ExtensionFormInfo.module.css
@@ -1,0 +1,12 @@
+.requiredText {
+  color: red;
+}
+
+.dateInputField {
+  padding-right: 2ch; /* Adjust space for icons or padding */
+  min-width: 10ch; /* Minimum width for empty input */
+}
+
+.dateInputContainer {
+  display: inline-block;
+}
